Move and consolidate event serializers
It was starting to get repetitive, so I felt it would be best for maintainability to define an object-oriented inheritance structure for the sake of DRY.
